End Home Screen and Update Screen

+++ app/src/main/java/com/example/areader/model/apiModel/ListPriceX.kt
@@ -1,6 +1,6 @@
 package com.example.areader.model.apiModel
 
 data class ListPriceX(
-    val amountInMicros: Int,
+    val amountInMicros: Double,
     val currencyCode: String
 )
