+++ app/src/main/java/com/example/areader/model/apiModel/RetailPrice.kt
@@ -1,6 +1,6 @@
 package com.example.areader.model.apiModel
 
 data class RetailPrice(
-    val amountInMicros: Int,
+    val amountInMicros: Double,
     val currencyCode: String
 )
